refactor: deduplicate filterHeaders and createSampleRecord across storage adapters
