+++ server/storage/__tests__/helpers.ts
@@ -1,0 +1,29 @@
+import type { RequestRecord } from '../../../common/types';
+
+export const createSampleRecord = (
+  id: string,
+  bucket: string,
+  timestamp?: string,
+  headers?: Record<string, string>,
+): RequestRecord => ({
+  id,
+  timestamp: timestamp || new Date().toISOString(),
+  bucket,
+  request: {
+    method: 'POST',
+    protocol: 'http',
+    host: 'localhost',
+    port: 3000,
+    pathQuery: `/hook/${bucket}/test`,
+    path: `/hook/${bucket}/test`,
+    args: '/test',
+    queryString: '',
+    query: {},
+    headers: headers || {
+      'content-type': 'application/json',
+      'user-agent': 'test-agent',
+    },
+    bodyRaw: '{"test": "data"}',
+    bodyJson: { test: 'data' },
+  },
+});
