+++ server/storage/__tests__/memory.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from 'bun:test';
-import type { RequestRecord } from '../../../common/types';
 import { MemoryStorageAdapter } from '../memory';
+import { createSampleRecord } from './helpers';
 import { createStorageInterfaceTests } from './shared/storage-interface.test';
 
 // Run shared interface tests
@@ -11,34 +11,6 @@ createStorageInterfaceTests(
 
 describe('MemoryStorageAdapter - Specific Implementation', () => {
   let adapter: MemoryStorageAdapter;
-
-  const createSampleRecord = (
-    id: string,
-    bucket: string,
-    timestamp?: string,
-    headers?: Record<string, string>,
-  ): RequestRecord => ({
-    id,
-    timestamp: timestamp || new Date().toISOString(),
-    bucket,
-    request: {
-      method: 'POST',
-      protocol: 'http',
-      host: 'localhost',
-      port: 3000,
-      pathQuery: `/hook/${bucket}/test`,
-      path: `/hook/${bucket}/test`,
-      args: '/test',
-      queryString: '',
-      query: {},
-      headers: headers || {
-        'content-type': 'application/json',
-        'user-agent': 'test-agent',
-      },
-      bodyRaw: '{"test": "data"}',
-      bodyJson: { test: 'data' },
-    },
-  });
 
   beforeEach(() => {
     adapter = new MemoryStorageAdapter();
